Moves SuggestionData information into the scorecard package
