+++ capstone/src/main/java/com/google/univiz/SuggestionDataApi.java
@@ -1,5 +1,7 @@
 package com.google.univiz;
 
+import com.google.univiz.api.scorecard;
+
 /**
  * Communicates directly with the data source to get a list of college suggestions related to a
  * client query
